Refactor blog page to use getAllPosts with search functionality; implement PostCard component for post display; enhance header search handling.
